remove default beacon config from beacon
builders provide valid config
