+++ src/main/java/com/dynatrace/openkit/AppMonOpenKitBuilder.java
@@ -49,7 +49,7 @@ public class AppMonOpenKitBuilder extends AbstractOpenKitBuilder {
         BeaconCacheConfiguration beaconCacheConfiguration = new BeaconCacheConfiguration(getBeaconCacheMaxRecordAge(),
             getBeaconCacheLowerMemoryBoundary(),
             getBeaconCacheUpperMemoryBoundary());
-        BeaconConfiguration beaconConfiguration = new BeaconConfiguration(1,
+        BeaconConfiguration beaconConfiguration = new BeaconConfiguration(BeaconConfiguration.DEFAULT_MULITPLICITY,
             getDataCollectionLevel(),
             getCrashReportLevel());
         return new Configuration(
