+++ src/main/java/com/dynatrace/openkit/api/OpenKitConstants.java
@@ -16,9 +16,6 @@
 
 package com.dynatrace.openkit.api;
 
-import com.dynatrace.openkit.core.configuration.CrashReportingLevel;
-import com.dynatrace.openkit.core.configuration.DataCollectionLevel;
-
 /**
  * Defines constant values used in OpenKit
  */
