+++ src/test/java/com/dynatrace/openkit/core/ActionImplTest.java
@@ -22,8 +22,8 @@ import com.dynatrace.openkit.api.WebRequestTracer;
 import com.dynatrace.openkit.core.caching.BeaconCacheImpl;
 import com.dynatrace.openkit.core.configuration.BeaconConfiguration;
 import com.dynatrace.openkit.core.configuration.Configuration;
-import com.dynatrace.openkit.core.configuration.CrashReportingLevel;
-import com.dynatrace.openkit.core.configuration.DataCollectionLevel;
+import com.dynatrace.openkit.CrashReportingLevel;
+import com.dynatrace.openkit.DataCollectionLevel;
 import com.dynatrace.openkit.protocol.Beacon;
 import com.dynatrace.openkit.providers.ThreadIDProvider;
 import com.dynatrace.openkit.providers.TimingProvider;
